+++ src/main.py
@@ -1,0 +1,19 @@
+DEBUG = True
+
+
+def main(debug_mode):
+    """
+    Summary: 
+        Entry point of program. 
+        If debug mode then test code is run, otherwise
+        main program is run
+
+    Params:
+        Takes global debug as bool
+
+    """
+    print("Hello")
+
+
+if __name__ == '__main__':
+    main(DEBUG)
